Add model new and recode typo
