+++ database/visitsUp.js
@@ -1,8 +1,8 @@
-const siteViews = require('./model/visits');
+const { visits } = require('./model/index');
 
 const siteViewsUp = function () {
-	siteViews
-		.findByIdAndUpdate('62e1556682dc251b0d97ad7e', { $inc: { counter: 1 } }, { new: true })
+	visits
+		.findByIdAndUpdate('62e1556682dc251b0d97ad7c', { $inc: { counter: 1 } }, { new: true })
 		.then((data) => {
 			console.log(data.counter);
 		})
